fix size of bufferPool and sender chans for custom writers
